application/requirements.txt
blinker==1.6.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
